@@ -4,7 +4,7 @@ author: Pavel <pavelgj@gmail.com>
 description: A Dart port of https://github.com/qiao/PathFinding.js library
 homepage: https://github.com/pavelgj/PathFinding.dart
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: ">=2.16.1 <3.0.0"
 dev_dependencies:
   benchmark_harness: any
   test: any
