@@ -23,11 +23,14 @@ import '../core/heuristic.dart';
 class AStarFinder {
   bool allowDiagonal;
   bool dontCrossCorners;
-  HeuristicFn heuristic;
+  HeuristicFn? heuristic;
   int weight;
 
-  AStarFinder({this.allowDiagonal: false, this.dontCrossCorners: false,
-      HeuristicFn heuristic, this.weight: 1}) {
+  AStarFinder(
+      {this.allowDiagonal: false,
+      this.dontCrossCorners: false,
+      HeuristicFn? heuristic,
+      this.weight: 1}) {
     this.heuristic = heuristic == null ? Heuristic.manhattan : heuristic;
   }
 
@@ -38,15 +41,22 @@ class AStarFinder {
    */
   findPath(startX, startY, endX, endY, Grid grid) {
     var openList = new Heap((nodeA, nodeB) {
-      return nodeA.f - nodeB.f;
-    }),
-    startNode = grid.getNodeAt(startX, startY),
-    endNode = grid.getNodeAt(endX, endY),
-    heuristic = this.heuristic,
-    allowDiagonal = this.allowDiagonal,
-    dontCrossCorners = this.dontCrossCorners,
-    weight = this.weight,
-    node, neighbors, neighbor, i, l, x, y, ng;
+          return nodeA.f - nodeB.f;
+        }),
+        startNode = grid.getNodeAt(startX, startY),
+        endNode = grid.getNodeAt(endX, endY),
+        heuristic = this.heuristic,
+        allowDiagonal = this.allowDiagonal,
+        dontCrossCorners = this.dontCrossCorners,
+        weight = this.weight,
+        node,
+        neighbors,
+        neighbor,
+        i,
+        l,
+        x,
+        y,
+        ng;
 
     // set the `g` and `f` value of the start node to be 0
     startNode.g = 0;
@@ -88,8 +98,9 @@ class AStarFinder {
         // can be reached with smaller cost from the current node
         if (neighbor.opened != true || ng < neighbor.g) {
           neighbor.g = ng;
-          neighbor.h = neighbor.h != null ?
-              neighbor.h : weight * heuristic(abs(x - endX), abs(y - endY));
+          neighbor.h = neighbor.h != null
+              ? neighbor.h
+              : weight * heuristic!(abs(x - endX), abs(y - endY));
           neighbor.f = neighbor.g + neighbor.h;
           neighbor.parent = node;
 
